+++ README.md
@@ -20,8 +20,8 @@
 ## step 3
 - [x] 킹을 잡으면 게임을 끝낸다.
 - [x] 킹이 잡힌 팀이 진다.
-- [ ] status를 입력하면 각 진영의 점수를 출력한다.
-- [ ] status를 입력하면 각 진영의 승패 결과를 출력한다. 
+- [x] status를 입력하면 각 진영의 점수를 출력한다.
+- [x] status를 입력하면 각 진영의 승패 결과를 출력한다. 
 - [x] 각 말은 점수를 갖는다. 
 - [ ] 리팩토링 
 
@@ -30,3 +30,4 @@
 - [x] 칸의 상태 확인하기 
 - [x] 칸의 말을 가져오기
 - [x] 칸이 어디에 위치해있는지 (좌표?)
+
