+++ src/main/java/chess/controller/ChessController.java
@@ -1,8 +1,8 @@
 package chess.controller;
 
-import chess.domain.BoardInitializer;
+import chess.domain.utils.BoardInitializer;
 import chess.domain.ChessGame;
-import chess.domain.Command;
+import chess.domain.board.Command;
 import chess.view.InputView;
 import chess.view.OutputView;
 
@@ -10,6 +10,16 @@ public class ChessController {
 
     public static final int COMMAND_INDEX = 0;
     public static final String SPACE = " ";
+
+    public void run() {
+        ChessGame chessGame = new ChessGame();
+        OutputView.printCommandInfo();
+        while (chessGame.isRunning()) {
+            String inputCmd = InputView.inputCommand();
+            Command command = Command.of(splitCommand(inputCmd));
+            command.apply(chessGame, inputCmd);
+        }
+    }
 
     public static void start(ChessGame chessGame, String command) {
         chessGame.initBoard(BoardInitializer.init());
@@ -22,9 +32,7 @@ public class ChessController {
         }
         chessGame.move(command);
         OutputView.printBoard(chessGame.getBoard(), chessGame.getTurn());
-        if (chessGame.isEnd()) {
-            OutputView.printTeamWin(chessGame.getWinTeam());
-        }
+        confirmKingDead(chessGame);
     }
 
     public static void end(ChessGame chessGame, String command) {
@@ -38,17 +46,13 @@ public class ChessController {
         OutputView.printStatus(chessGame.calculatePoint());
     }
 
-    public void run() {
-        ChessGame chessGame = new ChessGame();
-        OutputView.printCommandInfo();
-        while (chessGame.isRunning()) {
-            String inputCmd = InputView.inputCommand();
-            Command command = Command.of(splitCommand(inputCmd));
-            command.apply(chessGame, inputCmd);
-        }
-    }
-
     private String splitCommand(String inputCmd) {
         return inputCmd.split(SPACE)[COMMAND_INDEX];
+    }
+
+    private static void confirmKingDead(ChessGame chessGame) {
+        if (chessGame.isEnd()) {
+            OutputView.printTeamWin(chessGame.getWinTeam());
+        }
     }
 }
