+++ src/main/java/chess/domain/board/Command.java
@@ -1,6 +1,8 @@
-package chess.domain;
+package chess.domain.board;
 
 import chess.controller.ChessController;
+import chess.domain.ChessGame;
+
 import java.util.Arrays;
 import java.util.function.BiConsumer;
