+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,23 +1,16 @@
 package chess.domain.piece;
 
-import chess.domain.Direction;
-import chess.domain.Strategy;
-import chess.domain.Team;
+import chess.domain.board.Direction;
+import chess.domain.dto.Strategy;
+import chess.domain.board.Team;
 
-public class Bishop implements Piece {
+public class Bishop extends AbstractPiece{
 
     public static final double POINT = 3;
     private static final int MOVE_RANGE = 8;
 
-    private final Team team;
-
     public Bishop(Team team) {
-        this.team = team;
-    }
-
-    @Override
-    public Team getTeam() {
-        return team;
+        super(team);
     }
 
     @Override
@@ -27,22 +20,10 @@ public class Bishop implements Piece {
 
     @Override
     public String getName() {
-        if (team == Team.BLACK) {
+        if (isBlackTeam()) {
             return "B";
         }
         return "b";
-    }
-
-    @Override
-    public void checkTurn(Team team) {
-        if (this.team != team) {
-            throw new IllegalArgumentException("[ERROR] 상대 팀의 차례입니다.");
-        }
-    }
-
-    @Override
-    public boolean isSameTeam(Team team) {
-        return this.team == team;
     }
 
     @Override
