+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,10 +1,10 @@
 package chess.domain.piece;
 
-import chess.domain.Direction;
-import chess.domain.Strategy;
-import chess.domain.Team;
+import chess.domain.board.Direction;
+import chess.domain.dto.Strategy;
+import chess.domain.board.Team;
 
-public class Pawn implements Piece {
+public class Pawn extends AbstractPiece {
 
     public static final double POINT = 1;
     public static final int BLACK_PAWN_START_LINE = 7;
@@ -12,15 +12,13 @@ public class Pawn implements Piece {
     public static final int MOVE_FIRST_RANGE = 2;
     public static final int MOVE_DEFAULT_RANGE = 1;
 
-    private final Team team;
-
     public Pawn(Team team) {
-        this.team = team;
+        super(team);
     }
 
     @Override
     public Strategy strategy() {
-        if (team == Team.BLACK) {
+        if (isBlackTeam()) {
             return new Strategy(Direction.blackPawnDirection(), MOVE_FIRST_RANGE);
         }
         return new Strategy(Direction.whitePawnDirection(), MOVE_FIRST_RANGE);
@@ -28,27 +26,10 @@ public class Pawn implements Piece {
 
     @Override
     public String getName() {
-        if (team == Team.BLACK) {
+        if (isBlackTeam()) {
             return "P";
         }
         return "p";
-    }
-
-    @Override
-    public Team getTeam() {
-        return team;
-    }
-
-    @Override
-    public void checkTurn(Team team) {
-        if (this.team != team) {
-            throw new IllegalArgumentException("[ERROR] 상대 팀의 차례입니다.");
-        }
-    }
-
-    @Override
-    public boolean isSameTeam(Team team) {
-        return this.team == team;
     }
 
     @Override
