+++ src/test/java/chess/domain/BoardInitializerTest.java
@@ -2,6 +2,7 @@ package chess.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import chess.domain.utils.BoardInitializer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
