+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import chess.domain.Team;
+import chess.domain.board.Team;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
